Testing with trace record for all agents

+++ src/main/java/rldevs4j/agents/ac/ContinuosACActor.java
@@ -3,5 +3,5 @@ package rldevs4j.agents.ac;
 import org.nd4j.linalg.api.ndarray.INDArray;
 
 public interface ContinuosACActor extends ACActor{
-    public double[] action(INDArray obs);
+    public float[] action(INDArray obs);
 }
